+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Updated TOC for patch 11.1.0
+- Added a minimap icon. Left-click to open settings, or right-click to get a context menu of options.
 [Full Changelog](https://github.com/Pr3vention/UniversalBar/compare/1.1.8...1.1.9)